Fixed FindNearestLocation, TrainView and TransitView not receiving the current GPS location of the phone.  NTA no longer crashes when selected start and end stops.  Reversing the stops causes the data to be immediately loaded.  Find a bug in the Comments form, used NSString when NSDate should have been used.  This is version 1.0.9.1 as submitted to Apple on January 20th, 2015.

+++ iSEPTAv2/Subclasses/GTFS/GTFSFunctions.h
@@ -25,7 +25,7 @@
 //-(UInt32) getServiceID;
 //-(UInt32) getServiceIDForRouteType:
 
--(UInt32) getServiceIDFor:(NSString*) date;
+-(NSInteger) getServiceIDFor:(NSString*) date;
 
 
 @end
